Extract dummy delegate bridge class for reuse
